dna:course: separation of module module into multiple files under name section
This commit message is a perfect example of Hedayat's case for renaming
module -> section. It removes the ambiguity when talking about the app logic
and app parts (such as Rust modules).
Since this commit only changes the structure I have avoided renaming
module -> section completely. This will be done later.