dna::course: section anchor-first refactoring
(This commit isn't split into smaller chunks to make sure that code
compiles for every commit.)
Refactoring of Section is done for anchor-first pattern: anchor for
entry is created first and then it's address is written into data entry.
This anchor entry is used as kind of ID of the Section entry because
it's address stays constant. Therefore, all links to/from Section are
defined for this anchor entry and Section entry is used to contain
latest version of the data.

Changes include:
- anchor_address field in Section
- anchor entry definition
- new logic in CRUD handlers for Section:
	- create now returns SectionAnchor address to be used to
reference this section everywhere else
	- update removes link to previous section entry, updates section
entry and creates a new link
	- delete only removes anchor because without it section entries
would be unreachable
- renaming all section_address -> section_anchor_address in all methods
accross the zome
- stub for delete_section handler in course::handlers