dna::course: Content handlers refactoring removing update loops
Now every time Content is updated, the corresponding link from SectionAnchor to
previous version of this Content entry is removed and new link to the latest Content
entry is created.
This would ensure that updates are always performed on a different address (of the
latest entry) and any agent node will still be able to reconstruct the actual order of
events regardless of the order in which it receives these updates.