dna::course: Course entry anchor-first refactoring
8/8 integration tests are passing!

Implementation of the anchor-first pattern for the Course entry.
Now every Course is split into two parts:
1) CourseAnchor that serves as it's ID: a never changing entry that can
only be create/removed and is used to create any kind of links to/from
course
2) Course entry that contains latest course data.

Additionaly, this commit renames anchor that keeps track of all courses:
it's now CourseCatalogAnchor. It's a static anchor created once there's
a single course in the DNA and it keeps links to CourseAnchor entries
that correpond to particular courses.

Other changes include:
- anchor_address field in Course
- CourseAnchor entry definition
- new logic in CRUD handlers for Course that takes care of the
CourseAnchor -> Course relationship
- renaming of all course_address -> course_anchor_address to emphasize
that we're always operating with anchor address for any course-related
actions